Drop Next cache in workflow
Simplify the workflow, it will reduce maintenance burden.

It might be useful for SSR, but for SSG is more confusing than helpful, and we build anyhow in <1m

In particular, cache prevents pages regeneration, even when the source files are changed, since the latest cache were always restored, if packages were unchanged, and I suspect Next was not regenerating pages because the cached files have a recent timestamp, since they are created downloading the cache

Packages are still cached by the `actions/setup-node` action.